Some test format changes
